Use mc_ongoing instead of ~pe_ongoing
The code should be equivalent.  However, we have a ZCU104 board where
the old implementation, for poorly explained reasons, leads to an
inability to initialize memory.  The new implementation does not appear
to have this issue.  Just to get past this hurdle, we make a minor change
to the implementation.