scripts/spd_eeprom: parse only the first memory dump
When the bitstream has been compiled with module parameters from SPD,
then BIOS will compare the SPD used with the one present on the chip
when reading it. If there is a mismatch the reference memory will be
shown in the second dump. We should then parse only the first dump.