scripts: generate row_pairs with args checking
Moved row generation up, as there already is a tree of `if`s which
check `--row-pairs` choice.

Changed `--all-rows` loop to work on list of row pairs to be
consistent with other choices.

Added hammer-only branch as well to have an `if` expression that checks
all possible attack choices.

Now there will be a hard error, when no attack operation is provided
from the following list:
- `--hammer-only`
- `--all-rows`
- `--row-pairs sequential`
- `--row-pairs const`
- `--row-pairs random`

Signed-off-by: Michal Sieron <msieron@antmicro.com>