Update board defines in the Makefile
Top name for zcu104 is xilinx_zcu104
Remove ddr5_tester target as DDR5 is not merged yet

Unindent TOP, OFL_BOARD assignments as well as an error if given target
is unsupported.
Even though assignments can be indented and still work,
the `$(error ...)` cannot be. If it is, `make` thinks it's a part of
a recipe and errors with `recipe commences before first target`.

Signed-off-by: Michal Sieron <msieron@antmicro.com>