scripts/spd_eeprom: remove timeout for ddrctrl_init_done
In the current script, when the `mem_timeout` would run out, the script
would just go to the `read_spd` function, as if memory init was done.
That is incorrect.

Instead, wait until the `ddrctrl_init_done` is set regardless how long
it would take.

I also increased the timeout for pexpect to 30 s as 6 s might be too low.

Signed-off-by: Michal Sieron <msieron@antmicro.com>