Manually add clock constraints to the eth pads
S7PHY uses RX clock to generate both RX and TX clocks.
Default LiteX constraints are attached to the RX and TX
clock nets, which works with Vivado 2020.2, but causes
Critical warnings in the Viviado 2023.2.
This commit disables default constraint generation and
creates clock constraint using ethernet clock pad.

As a side note Xilinx/AMD warns against creating clock
constraint on the generated clocks,
such as that sourced from the PLL,
which is the case for the eth_tx clock in the S7PHY.

Signed-off-by: Maciej Dudek <mdudek@antmicro.com>